refactor: Divide arrayDivider file in three files and move to folder all update and newsletter controllers
