+++ env.dist
@@ -1,1 +1,21 @@
 YOUTUBE_API_KEY=api de youtube para tu app
+
+GOOGLE_GEMINI_API_KEY=
+
+EMAIL_SERVICE="gmail"
+EMAIL_PASS=""
+EMAIL_USER=""
+
+# This was inserted by `prisma init`:
+# Environment variables declared in this file are automatically made available to Prisma.
+# See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
+
+# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
+# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
+
+DATABASE_URL=""
+
+JWT_SECRET=""
+
+JWT_EXPIRES_IN=
+
