+++ src/services/newsletter/createAINewsletter.service.ts
