+++ src/util/arrayDivider.ts
@@ -3,21 +3,3 @@ export function chunk<T>(arr: T[], size: number): T[][] {
     arr.slice(i * size, i * size + size)
   );
 }
-
-export function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-  return date.toLocaleDateString('en-US', {
-    month: 'short',
-    day: 'numeric',
-    year: 'numeric',
-  });
-}
-
-export function truncateText(text: string, maxLength: number = 120): string {
-  if (text.length <= maxLength) return text;
-  const truncated = text.substring(0, maxLength);
-  const lastSpace = truncated.lastIndexOf(' ');
-  return (
-    (lastSpace > 0 ? truncated.substring(0, lastSpace) : truncated) + '...'
-  );
-}
