+++ src/util/truncateText.ts
@@ -1,0 +1,8 @@
+export function truncateText(text: string, maxLength: number = 120): string {
+  if (text.length <= maxLength) return text;
+  const truncated = text.substring(0, maxLength);
+  const lastSpace = truncated.lastIndexOf(' ');
+  return (
+    (lastSpace > 0 ? truncated.substring(0, lastSpace) : truncated) + '...'
+  );
+}
